market-backend-api/requirements.txt
alembic
annotated-types
anyio
openpyxl
bcrypt
cffi
click
colorama
cryptography
dnspython
ecdsa
email-validator
fastapi
greenlet
h11
httptools
idna
Mako
MarkupSafe
motor
passlib
pyasn1
pycparser
pydantic
pydantic_core
pymongo
python-dateutil
python-dotenv
python-jose
python-multipart
PyYAML
rsa
six
sniffio
SQLAlchemy
starlette
typing-inspection
typing_extensions
uvicorn
watchfiles
websockets